Restructure - each file in modules/* will be associated with Store, Actions, Component that will be kept in the same file instead of different files as they are cohesive in nature.
