+++ app/modules/multiply.jsx
@@ -17,20 +17,71 @@
 var React = require('react');
 var Reflux = require('reflux');
 
-var multiplyStore = require('./multiply-store');
-var commandbox = require('../../actions/commandbox-actions.js');
+var commandbox = require('./commandbox');
 
-var Multiply = React.createClass({
-  mixins: [Reflux.connect(multiplyStore)],
+var Store = Reflux.createStore({
+  init() {
+    this.listenTo(commandbox.Store, this.onCommandUpdate);
+  },
+  onCommandUpdate(cmd) {
+    var obj = {
+      three: false,
+      nine: false,
+      twelve: false,
+      multiplyTotal: 0
+    };
+    if (!cmd.arg) {
+      cmd.arg = '';
+    }
+    cmd.arg.split(' ').map(function(s) {
+      switch(s) {
+      case 'three':
+        obj.three = true;
+        if (obj.multiplyTotal === 0) {
+          obj.multiplyTotal = 1
+        }
+        obj.multiplyTotal *= 3;
+        break;
+      case 'nine':
+        obj.nine = true;
+        if (obj.multiplyTotal === 0) {
+          obj.multiplyTotal = 1
+        }
+        obj.multiplyTotal *= 9;
+        break;
+      case 'twelve':
+        obj.twelve = true;
+        if (obj.multiplyTotal === 0) {
+          obj.multiplyTotal = 1
+        }
+        obj.multiplyTotal *= 12;
+        break;
+      }
+    })
+    this.trigger(obj);
+  },
+  data: {
+    three: false,
+    nine: false,
+    twelve: false,
+    mutliplyTotal: 0
+  },
+  getInitialState() {
+    return this.data
+  }
+});
+
+var Component = React.createClass({
+  mixins: [Reflux.connect(Store)],
   componentDidMount() {
-    commandbox.commandSet('multiply');
+    commandbox.Actions.commandSet('multiply');
   },
   multiplyHandler() {
     var argSet = '';
     if (this.state.three) argSet+='three ';
     if (this.state.nine) argSet+='nine ';
     if (this.state.twelve) argSet+='twelve ';
-    commandbox.argSet(argSet);
+    commandbox.Actions.argSet(argSet);
   },
   multiplyHandlerThree() {
     this.state.three = !this.state.three;
@@ -80,4 +131,4 @@ var Multiply = React.createClass({
   }
 });
 
-module.exports = Multiply;
+module.exports = {Component};
